upd ai docs branch

+++ pyreason/scripts/interval/interval.py
@@ -64,8 +64,8 @@ class Interval(structref.StructRefProxy):
         lower = max(self.lower, interval.lower)
         upper = min(self.upper, interval.upper)
         if lower > upper:
-            lower = np.float32(0)
-            upper = np.float32(1)
+            lower = np.float64(0)
+            upper = np.float64(1)
         return Interval(lower, upper, False, self.lower, self.upper)
 
     def to_str(self):
